Web scraper successfully able to scrape game info
"The web scraper stores basic information about each free-to-play game it scrapes, such as the: name, price, description. It is also given a user friendly ID and a UUID. An image of the game is also stored locally. Cookies are accepted and age restricted games are ignored when scraping occurs."